WinVersion: Use NativeArchitecture and add process Architecture
Get details on the architecture the process is running as (Architecture), and the native processor architecture Windows is running on (NativeArchitecture). This uses IsWow64Process2() for ARM support, and if not available, GetNativeSystemInfo(), GetSystemInfo() and IsWow64Process().

The previous implementation would only work on AMD64 when using WoW for x86, but would not work as expected in IA64. This fixes that the native architecture is correctly provided.

Issue: DOTNET-397