NetVersion: Properly report the .NET version if no direct match
In case that the version of .NET that is installed is a preview version, which is unknown, then choose the version n the middle, not the next version higher. For example, on Windows 11 preview the version is 533320, but the documented release version is 533325 for .NET 4.8.1. Prior to this commit, the version was being reported as 4.8.1, but the release version detected is actually "older", so it should be reported as 4.8 or later, not 4.8.1.

Issue: DOTNET-659