NetVersion: Provide the Target Version as a `Version` object
For .NET 4.0, we specifically look at the profile to determine the version of the framework that is used. On Windows XP, the version is 4.0.30319. But on computers with .NET 4.x installed, it bumps, even though we still print Version 4 Client / 4 Full, and then the release to print the proper version as a new object.

Issue: DOTNET-842