NetVersion: Document and provide examples
Review and document how to detect .NET versions. This leads to easier refactoring of the `NetVersion` class later with concrete details (such as registry dumps).

Tests were done on Windows XP and Windows 10/11.

We note that the Microsoft documentation for detecting .NET 1.0 is incorrect on Windows XP. We also note the current implementation for .NET 3.0 is incorrect in this repository, unless .NET 3.5 is installed.

Issue: DOTNET-843