NetVersion: Rework API for Framework and Install version
After implementing dbbc633 ("NetVersion: Provide the Target Version as a `Version` object), the API is now confusing.

One important use case is to compare framework versions with each other. This is now done with the `FrameworkVersion` property.

The `Version` property is now the `InstallVersion` property for what is installed and is no longer a string, as it was also partially provided by `Description`.

A new `VersionType` property is created so one can know which .NET framework (.NET Core, .NET Framework or Mono) is being used.

With the rework, a few bugs were fixed:

* Possible null pointer accesses for registry usage (although this could only happen if the registry key was removed after iterating, which might be weird and only during a deinstallation)
* Keys with versions are now case insensitive (v or V are accepted)

Issue: DOTNET-845