NetVersion: Only return one current runtime
It is confusing to return both the `Environment.Version` and the Mono runtime as a list. Instead, just return the current runtime as a single object.

The `MonoRuntime` was extended to find the path to the runtime and to return the `mscorlib` version on Windows. A common interface was provided to make it the same for the frameworks available and the running framework.

Issue: DOTNET-848, DOTNET-850