NetVersion: Use the current MonoRuntime to also determine the installed location
We search multiple possible paths, and while doing so, remember the location as given by the file system to detect duplicates (from RJCP.IO.Path namespace). This makes it easier to determine the location of the current mono runtime, and if the path was not found prior, it is also added as an available runtime.

The path is not checked as there is no implementation to check the architecture of the binary when found (if it's 64-bit or 32-bit).

Tested with:
* Mono 3.2.3 on Windows XP
* Mono 6.1.0.200 on Windows 7 x64

Issue: DOTNET-848, DOTNET-851