Net/Windows Version: Add references for OS Compatibility
The `Platform` methods are now marked with `[SupportedOSPlatformGuard]` so that the diagnostic works properly on .NET 6.0 and later. When adding this a breaking change was made, to *remove* OSX as being supported (so that the supported platform is Linux). I do not have a MacOS-X platform to test with and code is likely broken all over the place anyway.

Suppress instances of `warning CA1416: This call site is reachable on all platforms`. Methods are given this attribute, and then the diagnostic executes the callchain, so the location of the calls can be properly audited.

Issue: DOTNET-938