WinVersionApiQuery: Native tool for querying Windows API on WinXP SP3 - Win11
Provide a tool (that works on Windows XP, Vista, 7, 8, 8.1, 10, 11) that queries the various APIs and writes it to the console (so it can be stored to a file). Solution file is Visual Studio 2022 (17.14.2).

Use the tool without an argument to print to the console. Provide a filename as the only argument to write to a file. Writing to a file writes in the Operating Systems native encoding (e.g. on Windows 10 it writes the file in Unicode, so that branding strings are properly represented).

Issue: DOTNET-1045, DOTNET-1047