WinVersionApiQuery: Refactor sources
Move the source code to a common directory. This allows us to create new projects for older versions of Visual Studio, that we'll use to build WinXP compatible binaries.

This removes the deprecation notice of Visual Studio 2022.

We can have older projects using the same sources in parallel.

Issue: DOTNET-1051