WinVersionApiQuery: Provide WinXP (VS2005) build solution
The Visual Studio 2005 solution is run on Windows XP SP3 and compiles for 32-bit and 64-bit. It shares the same sources as the more modern Visual Studio 2022 version. This should work also on Win 2000.

The output of VS2005 generated binary is the same as the VS2022 generated binary, tested on WinXP and Windows 10 x64.

Issue: DOTNET-1051