WinVersionApiQuery: Build on Windows NT 4.0 SP6a
The x86 release binary now links the runtime, rather than a DLL, to avoid the use of a symbol that is not exported on Windows NT 4.0. Now the 32-bit binary will run on Windows NT 4.0 up to the most modern version of Windows.

To make it easier to port, the "win32_compat.h" file is renamed to "stdafx.h" which is included first in the sources. Then we can set the `WINVER` version in one place to compile for different targets if needed for testing.

Issue: DOTNET-1051