WinVersionApiQuery: Compile against Visual Studio 2012
It is possible to get the Win10.v143 binaries working on Vista SP1 and earlier, but this requires copying system files since the Kernel in SP2 has been refactored to have API libraries, like `api-ms-win-core-console-l1-1-0.dll`. The VS2012 binaries don't need this and are simpler to get working.

The binaries will run on Windows XP SP3 after the Visual Studio 2012 Redistributable Packages are installed. It won't run on Windows XP (SP0).

Issue: DOTNET-1058