WinVersion: Rework Windows 9x
Remove WinProductType.Win9x as this is not defined by Windows and not needed, nor used.

Windows95a can be determined via the registry, with the "SubVersionNumber", but it isn't in the CSDVersion. So emulate it.

There is no way via the API or registry to determine the difference between Windows 95 OSR2 and Windows 95 OSR2.1 (with USB support). These appear the same (except for a USB driver).

Issue: DOTNET-1059